+++ HashFunction/Murmurhash3New.h
@@ -3,4 +3,4 @@
 
 static inline uint64_t getblock64(const uint64_t *p, int i);
 static inline uint64_t fmix64(uint64_t k);
-unsigned long long MurmurHash3_x64_64(const void *key, int len, uint32_t seed);
+uint64_t MurmurHash3_x64_64(const void *key, int len, uint32_t seed);
