+++ KVSSD.h
@@ -6,10 +6,11 @@
 #include <stdlib.h>
 #include <stdio.h>
 #include <time.h>
+#include <stdint.h>
 
 typedef struct {
     int threshold;
-    unsigned long long capacity;
+    uint64_t capacity;
     int page_size;
     int pages_per_block;
     int block_size;
@@ -32,10 +33,10 @@ typedef struct {
 } KVSSD;
 
 // Function Prototypes
-void init_KVSSD(KVSSD *ssd, unsigned long long capacity, int page_size, int slab_size, int threshold);
+void init_KVSSD(KVSSD *ssd, uint64_t capacity, int page_size, int slab_size, int threshold);
 int gmd_size(KVSSD *kvssd);
-unsigned long long hash_k(const char *key);
-int get_translation_page(KVSSD *ssd, unsigned long long key_hash);
+uint64_t hash_k(const char *key);
+int get_translation_page(KVSSD *ssd, uint64_t key_hash);
 bool write(KVSSD *kvssd, const char *key, int klen, int val, int vlen);
 bool read(KVSSD *kvssd, const char *key);
 bool delete(KVSSD *kvssd, const char *key);
